requirements.txt
asgiref==3.6.0
base58==2.1.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2
docopt==0.6.2
ecdsa==0.18.0
idna==3.4
mysql-connector-python==8.0.33
mysqlclient==2.1.1
pipreqs==0.4.13
protobuf==3.20.3
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqt5-tools==5.15.9.3.3
PyQt5Designer==5.14.1
python-dotenv==1.0.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.29.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.2
urllib3==1.26.15
yarg==0.1.9
